Initial commit of 1 port USB relay firmware
This is a basic set of code which has the same functionality as the
default firmware on the www.dcttech.com 1 port USB Relay. There are
several points where there's a lack of checking that probably should
be there, but the original firmware lacks this too. Should be fixed up.
Also support for the boards with more relays would be nice, but I don't
have any to play with.